initial commit
this initial commit doesn't have basic functionality of the create command working yet. running into issues with Docker Machine creating errors when there shouldn't be during the net.DialTimeout command. Docker Machine doesn't wait and kills the process with an 'unexpected EOF' error. However, this works fine inside of a script. Lastly, and unfortunately, I'm going to need some real hardware to test this. The Vagrant environment won't be suitable because of the inability to have SSH access into the private network. I may be able to use my vSphere homelab but will need to research the possibility of sticking it on its own VLAN and hoping to give my laptop access to it. That way I'm not moving binaries for testing every 30 seconds

Signed-off-by: Kendrick Coleman <kendrickcoleman@gmail.com>