Add Glide file and handle vendored bindings
gorackhd and gorackhd-redfish are generated by different versions
of go-swagger. gorackhd is newer, but that same (stable) version of
go-swagger does not produce compilable code for gorackhd-redfish.

Deal with this for now by importing the two different/needed versions
of strfmt and httpkit. This should be temporary and be revisited when
go-swagger 0.6.0 comes about.