Move chooseNode() to PreCreateCheck()
When a new host is provisioned, the driver info is marshalled to
JSON and saved to disk. However, when calling Create(), any changes
made to the Driver struct are not persisted if we return an error.

In our case, this means that if we pick a node from RackHD using a
SKU (thereby picking a Node ID), but then dont' successfully provision
that node (for any reason), the Node ID is not persistent. A subsequent
call to docker-machine won't have this Node ID anymore, which means
that we can't properly delete it.

Work around this by doing the "pick a node from the pool" part in
the PreCreateCheck() instead.