Add --rackhd-sku-name option
The IDs in RackHD are not (meant to be) human-friendly, as they
are long hex strings. If the user knows a more usable name for
the SKU, allow them to use that instead.

This option is mutually exclusive with sku-id, and both sku-id and
sku-name are mutually exclusive with node-id.