Add --rackhd-worfklow-name support
When the workflow name is given (e.g. "Graph.InstallCoreOS"), that
workflow will be set on the node. The driver will then wait for
the workflow to finish before moving on.

The intent of the flag is to allow for OS installation. The timeouts
associated with waiting for the workflow to finish are currently
hardcoded to waiting for 60 minutes, and polled every 15 seconds.