Add initial SSH connect attempt retries
After we get an IP address for the node, we check to see if we can
reach the node on the IP:port for SSH. However in some cases the
node will not be ready yet. A specific example is the current state
of installing CoreOS. The workflow is finished as soon as the install
is finished and starts a reboot. Our logic would see that then attempt
to connect, which of course would fail because the node is rebooting.

Other OS installs have a "validate SSH" task in the workflows, which
doesn't show the workflow as finished until SSH into the node is
successful, but that isn't in place yet for CoreOS.

So, we add retries and sleep in between. I initially put the sleep
in net.DialTimeout(), but that is problematic because when the OS
first is coming up, there is a period where the TCP connection is
refused (instead of timing out). This is the period between when
the OS gets its networking up and when the SSH daemon starts. If
we don't sleep separately from net.DialTimeout(), we just go into a
tight loop with net.DialTimeout() returning immediately each time and
exhausting all our retries, instead of the desired effect of "connect
didn't work, wait then try again".