Use gorackhd v0.1.0 and add glide.lock
We should always be able to rebuild the exact same binary when we
tag a release. If we aren't checking in glide.lock we can't do that
because "glide up" may produce an updated dependency if it's not
pinned down in some other fashion.

Developers should be able to "glide up" as they need to, but to build
a release, it should be as simple as clone, glide install, and go build.