Update glide to lock down go-swagger
When using newer Glide (0.12), it picks up an error due to the fact
that gorackhd and gorackhd-redfish are using different versions of
go-swagger, a problem I have tried to resolve in the past but
unsuccessfully so far.

We can fix the error by locking down go-swagger to the older version
used by gorackhd-redfish. This is okay because only gorackhd-redfish is
using additional libraries that were in go-swagger but have since been
moved out to go-openapi.